+++ _examples/view/template_pug_3/bindata.go
@@ -215,9 +215,9 @@ type bintree struct {
 }
 
 var _bintree = &bintree{nil, map[string]*bintree{
-	"templates": &bintree{nil, map[string]*bintree{
-		"index.pug":  &bintree{templatesIndexPug, map[string]*bintree{}},
-		"layout.pug": &bintree{templatesLayoutPug, map[string]*bintree{}},
+	"templates": {nil, map[string]*bintree{
+		"index.pug":  {templatesIndexPug, map[string]*bintree{}},
+		"layout.pug": {templatesLayoutPug, map[string]*bintree{}},
 	}},
 }}