+++ hero/handler.go
@@ -11,31 +11,17 @@ import (
 	"github.com/kataras/golog"
 )
 
-var contextTyp = reflect.TypeOf((*context.Context)(nil)).Elem()
+// var genericFuncTyp = reflect.TypeOf(func(context.Context) reflect.Value { return reflect.Value{} })
 
-// IsContext returns true if the "inTyp" is a type of Context.
-func IsContext(inTyp reflect.Type) bool {
-	return inTyp.Implements(contextTyp)
-}
-
-// var genericFuncTyp = reflect.TypeOf(func(context.Context) interface{} { return nil })
-var genericFuncTyp = reflect.TypeOf(func(context.Context) reflect.Value { return reflect.Value{} })
-
-// IsGenericFunc reports whether the "inTyp" is a type of func(Context) interface{}.
-func IsGenericFunc(inTyp reflect.Type) bool {
-	return inTyp == genericFuncTyp
-}
+// // IsGenericFunc reports whether the "inTyp" is a type of func(Context) interface{}.
+// func IsGenericFunc(inTyp reflect.Type) bool {
+// 	return inTyp == genericFuncTyp
+// }
 
 // checks if "handler" is context.Handler: func(context.Context).
 func isContextHandler(handler interface{}) (context.Handler, bool) {
-	h, is := handler.(context.Handler)
-	if !is {
-		fh, is := handler.(func(context.Context))
-		if is {
-			return fh, is
-		}
-	}
-	return h, is
+	h, ok := handler.(context.Handler)
+	return h, ok
 }
 
 func validateHandler(handler interface{}) error {
@@ -72,7 +58,7 @@ func makeHandler(handler interface{}, errorHandler di.ErrorHandler, values ...re
 	}
 
 	funcInjector := di.Func(fn, values...)
-	funcInjector.ErrorHandler(errorHandler)
+	funcInjector.ErrorHandler = errorHandler
 
 	valid := funcInjector.Length == n